refactor: 定義の移動
