+++ server/infra/resource/src/records.rs
@@ -4,13 +4,12 @@ use chrono::{DateTime, Utc};
 use domain::{
     form::{
         answer::models::{AnswerAuthor, AnswerEntry, AnswerLabel, AnswerTitle, FormAnswerContent},
-        answer_entry_set::models::{AnswerSettings, DefaultAnswerTitle, ResponsePeriod},
         comment::models::{Comment, CommentContent},
         message::models::Message,
         models::{
-            ActiveForm, ArchivedForm, FormDescription, FormId, FormLabel, FormLabelId,
-            FormLabelIdSet, FormLabelName, FormMeta, FormSettings, FormTitle, QuestionSet,
-            WebhookUrl,
+            ActiveForm, AnswerSettings, ArchivedForm, DefaultAnswerTitle, FormDescription, FormId,
+            FormLabel, FormLabelId, FormLabelIdSet, FormLabelName, FormMeta, FormSettings,
+            FormTitle, QuestionSet, ResponsePeriod, WebhookUrl,
         },
         question::models::{Choice, Question, QuestionType},
     },
